refactor(examples) Update XGBoost comprehensive example (#4234)
