+++ examples/xgboost-comprehensive/pyproject.toml
@@ -1,15 +1,55 @@
 [build-system]
-requires = ["poetry-core>=1.4.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
-[tool.poetry]
-name = "xgboost-comprehensive"
-version = "0.1.0"
-description = "Federated XGBoost with Flower (comprehensive)"
-authors = ["The Flower Authors <hello@flower.ai>"]
+[project]
+name = "xgboost_comprehensive"
+version = "1.0.0"
+description = "Federated Learning with XGBoost and Flower (Comprehensive Example)"
+license = "Apache-2.0"
+dependencies = [
+    "flwr[simulation]>=1.11.1",
+    "flwr-datasets>=0.3.0",
+    "xgboost>=2.0.0",
+]
 
-[tool.poetry.dependencies]
-python = ">=3.9,<3.11"
-flwr = { extras = ["simulation"], version = ">=1.7.0,<2.0" }
-flwr-datasets = ">=0.2.0,<1.0.0"
-xgboost = ">=2.0.0,<3.0.0"
+[tool.hatch.build.targets.wheel]
+packages = ["."]
+
+[tool.flwr.app]
+publisher = "flwrlabs"
+
+[tool.flwr.app.components]
+serverapp = "xgboost_comprehensive.server_app:app"
+clientapp = "xgboost_comprehensive.client_app:app"
+
+[tool.flwr.app.config]
+# ServerApp
+train-method = "bagging"  # Choose from [bagging, cyclic]
+num-server-rounds = 3
+fraction-fit = 1.0
+fraction-evaluate = 1.0
+centralised-eval = false
+
+# ClientApp
+partitioner-type = "uniform"  # Choose from [uniform, linear, square, exponential]
+test-fraction = 0.2
+seed = 42
+centralised-eval-client = false
+local-epochs = 1
+scaled-lr = false
+params.objective = "binary:logistic"
+params.eta = 0.1  # Learning rate
+params.max-depth = 8
+params.eval-metric = "auc"
+params.nthread = 16
+params.num-parallel-tree = 1
+params.subsample = 1
+params.tree-method = "hist"
+
+[tool.flwr.federations]
+default = "local-simulation"
+
+[tool.flwr.federations.local-simulation]
+options.num-supernodes = 5
+options.backend.client-resources.num-cpus = 2
