--- examples/xgboost-comprehensive/requirements.txt
@@ -1,3 +1,0 @@
-flwr[simulation]>=1.7.0, <2.0
-flwr-datasets>=0.2.0, <1.0.0
-xgboost>=2.0.0, <3.0.0
