+++ examples/xgboost-comprehensive/xgboost_comprehensive/__init__.py
@@ -1,0 +1,1 @@
+"""xgboost_comprehensive: A Flower / XGBoost app."""
